scripts/requirements.txt
requests==2.32.5
WooCommerce==3.0.0
Markdown==3.5.2
python-dotenv==1.0.1
